requirements.txt
attrs==19.3.0
Automat==0.8.0
blinker==1.6.2
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
cloud-init==19.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
Flask==2.3.2
gunicorn==20.1.0
httplib2==0.14.0
huaweicloud-sdk-python-obs==3.21.8
huaweicloudsdkall==3.1.40
huaweicloudsdkantiddos==3.1.40
huaweicloudsdkaom==3.1.40
huaweicloudsdkaos==3.1.40
huaweicloudsdkapig==3.1.40
huaweicloudsdkapm==3.1.40
huaweicloudsdkas==3.1.40
huaweicloudsdkbcs==3.1.40
huaweicloudsdkbms==3.1.40
huaweicloudsdkbss==3.1.40
huaweicloudsdkbssintl==3.1.40
huaweicloudsdkcae==3.1.40
huaweicloudsdkcampusgo==3.1.40
huaweicloudsdkcbh==3.1.40
huaweicloudsdkcbr==3.1.40
huaweicloudsdkcbs==3.1.40
huaweicloudsdkcc==3.1.40
huaweicloudsdkcce==3.1.40
huaweicloudsdkccm==3.1.40
huaweicloudsdkcdm==3.1.40
huaweicloudsdkcdn==3.1.40
huaweicloudsdkces==3.1.40
huaweicloudsdkcfw==3.1.40
huaweicloudsdkcgs==3.1.40
huaweicloudsdkclassroom==3.1.40
huaweicloudsdkclouddeploy==3.1.40
huaweicloudsdkcloudide==3.1.40
huaweicloudsdkcloudpipeline==3.1.40
huaweicloudsdkcloudrtc==3.1.40
huaweicloudsdkcloudtable==3.1.40
huaweicloudsdkcloudtest==3.1.40
huaweicloudsdkcodeartsartifact==3.1.40
huaweicloudsdkcodeartsbuild==3.1.40
huaweicloudsdkcodecheck==3.1.40
huaweicloudsdkcodecraft==3.1.40
huaweicloudsdkcodehub==3.1.40
huaweicloudsdkcore==3.1.40
huaweicloudsdkcph==3.1.40
huaweicloudsdkcpts==3.1.40
huaweicloudsdkcse==3.1.40
huaweicloudsdkcsms==3.1.40
huaweicloudsdkcss==3.1.40
huaweicloudsdkcts==3.1.40
huaweicloudsdkdas==3.1.40
huaweicloudsdkdbss==3.1.40
huaweicloudsdkdc==3.1.40
huaweicloudsdkdcs==3.1.40
huaweicloudsdkddm==3.1.40
huaweicloudsdkdds==3.1.40
huaweicloudsdkdeh==3.1.40
huaweicloudsdkdevsecurity==3.1.40
huaweicloudsdkdevstar==3.1.40
huaweicloudsdkdgc==3.1.40
huaweicloudsdkdlf==3.1.40
huaweicloudsdkdli==3.1.40
huaweicloudsdkdms==3.1.40
huaweicloudsdkdns==3.1.40
huaweicloudsdkdris==3.1.40
huaweicloudsdkdrs==3.1.40
huaweicloudsdkdsc==3.1.40
huaweicloudsdkdwr==3.1.40
huaweicloudsdkdws==3.1.40
huaweicloudsdkecs==3.1.40
huaweicloudsdkeg==3.1.40
huaweicloudsdkeihealth==3.1.40
huaweicloudsdkeip==3.1.40
huaweicloudsdkelb==3.1.40
huaweicloudsdkeps==3.1.40
huaweicloudsdker==3.1.40
huaweicloudsdkevs==3.1.40
huaweicloudsdkfrs==3.1.40
huaweicloudsdkfunctiongraph==3.1.40
huaweicloudsdkga==3.1.40
huaweicloudsdkgaussdb==3.1.40
huaweicloudsdkgaussdbfornosql==3.1.40
huaweicloudsdkgaussdbforopengauss==3.1.40
huaweicloudsdkges==3.1.40
huaweicloudsdkgsl==3.1.40
huaweicloudsdkhilens==3.1.40
huaweicloudsdkhss==3.1.40
huaweicloudsdkiam==3.1.40
huaweicloudsdkiec==3.1.40
huaweicloudsdkief==3.1.40
huaweicloudsdkies==3.1.40
huaweicloudsdkimage==3.1.40
huaweicloudsdkimagesearch==3.1.40
huaweicloudsdkims==3.1.40
huaweicloudsdkiotanalytics==3.1.40
huaweicloudsdkiotda==3.1.40
huaweicloudsdkiotedge==3.1.40
huaweicloudsdkivs==3.1.40
huaweicloudsdkkafka==3.1.40
huaweicloudsdkkms==3.1.40
huaweicloudsdkkps==3.1.40
huaweicloudsdklakeformation==3.1.40
huaweicloudsdklive==3.1.40
huaweicloudsdklts==3.1.40
huaweicloudsdkmapds==3.1.40
huaweicloudsdkmas==3.1.40
huaweicloudsdkmeeting==3.1.40
huaweicloudsdkmetastudio==3.1.40
huaweicloudsdkmoderation==3.1.40
huaweicloudsdkmpc==3.1.40
huaweicloudsdkmrs==3.1.40
huaweicloudsdkmsgsms==3.1.40
huaweicloudsdknat==3.1.40
huaweicloudsdknlp==3.1.40
huaweicloudsdkocr==3.1.40
huaweicloudsdkoms==3.1.40
huaweicloudsdkorganizations==3.1.40
huaweicloudsdkosm==3.1.40
huaweicloudsdkprojectman==3.1.40
huaweicloudsdkrabbitmq==3.1.40
huaweicloudsdkram==3.1.40
huaweicloudsdkrds==3.1.40
huaweicloudsdkres==3.1.40
huaweicloudsdkrms==3.1.40
huaweicloudsdkrocketmq==3.1.40
huaweicloudsdkroma==3.1.40
huaweicloudsdksa==3.1.40
huaweicloudsdkscm==3.1.40
huaweicloudsdksdrs==3.1.40
huaweicloudsdksecmaster==3.1.40
huaweicloudsdkservicestage==3.1.40
huaweicloudsdksfsturbo==3.1.40
huaweicloudsdksis==3.1.40
huaweicloudsdksmn==3.1.40
huaweicloudsdksms==3.1.40
huaweicloudsdkswr==3.1.40
huaweicloudsdktms==3.1.40
huaweicloudsdkugo==3.1.40
huaweicloudsdkvas==3.1.40
huaweicloudsdkvcm==3.1.40
huaweicloudsdkvod==3.1.40
huaweicloudsdkvpc==3.1.40
huaweicloudsdkvpcep==3.1.40
huaweicloudsdkvss==3.1.40
huaweicloudsdkwaf==3.1.40
huaweicloudsdkworkspace==3.1.40
hyperlink==19.0.0
idna==2.8
importlib-metadata==6.6.0
importlib-resources==5.4.0
incremental==16.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.2
jsonschema==4.4.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.0.1
netifaces==0.10.4
oauthlib==3.1.0
pexpect==4.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycryptodome==3.10.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.7
python-debian===0.1.36ubuntu1
PyYAML==6.0
requests==2.22.0
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.19.1
six==1.16.0
sos==4.2
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
typing-extensions==4.6.0
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
Werkzeug==2.3.4
zipp==3.7.0
zope.interface==4.7.1
